+++ tests/codec/string.test.js
@@ -209,3 +209,46 @@ test("bare.writeString", (t) => {
         ]
     )
 })
+
+test("bare.writeFixedString", (t) => {
+    let bc = fromConfigBytes({ textEncoderThreshold: 0 })
+    bare.writeFixedString(bc, "bare")
+    t.deepEqual(toBytes(bc), [98, 97, 114, 101], "can natively write ASCII")
+
+    bc = fromBytes()
+    bare.writeFixedString(bc, "bare")
+    t.deepEqual(toBytes(bc), [98, 97, 114, 101], "can write ASCII")
+
+    bc = fromBytes()
+    bare.writeFixedString(bc, "éàù")
+    t.deepEqual(
+        toBytes(bc),
+        [195, 169, 195, 160, 195, 185],
+        "can write 2-bytes chars"
+    )
+
+    bc = fromBytes()
+    bare.writeFixedString(bc, "🎈🏃🏿‍♂️")
+    t.deepEqual(
+        toBytes(bc),
+        [
+            240, 159, 142, 136, 240, 159, 143, 131, 240, 159, 143, 191, 226,
+            128, 141, 226, 153, 130, 239, 184, 143,
+        ],
+        "can write surrogate chars"
+    )
+
+    bc = fromBytes()
+    bare.writeFixedString(
+        bc,
+        "\u3053\u3093\u306B\u3061\u306F\u3001\u4E16\u754C\uFF01"
+    )
+    t.deepEqual(
+        toBytes(bc),
+        [
+            0xe3, 0x81, 0x93, 0xe3, 0x82, 0x93, 0xe3, 0x81, 0xab, 0xe3, 0x81,
+            0xa1, 0xe3, 0x81, 0xaf, 0xe3, 0x80, 0x81, 0xe4, 0xb8, 0x96, 0xe7,
+            0x95, 0x8c, 0xef, 0xbc, 0x81,
+        ]
+    )
+})